requirements.txt
click==8.0.0
colorama==0.4.4
et-xmlfile==1.1.0
Flask==2.0.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
Werkzeug==2.0.0
